Re-edited, implemented and tested
